feat(player_wall_slide): Decompose `player.rs` file into `player_physics` and `player_animations`. Physics now handle only business logic but Animation supports only player animation
